third_party/blink/web_tests/http/tests/security/aboutBlank/security-context-grandchildren-alias.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
  <script>
    const document_domain_initial = "subdomain.example.test"
    const document_domain_new = "example.test";

    if (document.location.hostname != document_domain_initial) {
      document.location.hostname = document_domain_initial;
    } else {
      async_test(test => {
        const iframe = document.createElement('iframe');
        iframe.src = 'resources/iframe-with-about-blank-children.html';
        iframe.onload = test.step_func_done(() => {
          const doc0 = frames[0].frames[0].document;
          const doc1 = frames[0].frames[1].document;

          assert_equals(doc0.domain, document_domain_initial);
          assert_equals(doc1.domain, document_domain_initial);

          doc0.open();
          doc1.open();
          assert_equals(doc0.domain, document_domain_initial);
          assert_equals(doc1.domain, document_domain_initial);

          document.domain = document_domain_new;
          assert_equals(doc0.domain, document_domain_initial);
          assert_equals(doc1.domain, document_domain_initial);

          doc0.close();
          doc1.close();
          assert_equals(doc0.domain, document_domain_initial);
          assert_equals(doc1.domain, document_domain_initial);
        });
        document.body.appendChild(iframe);
      }, "document.domain behavior of grandchildren 'about:blank' " +
         "matches grandparent.");
    }
  </script>
</body>
</html>
